:sparkles: (EtuUTTApiTestCase) Added createUser method
Added createUser method. The methods takes 4 parameters : $firstName, $lastName, $login and $role. Role is not mendatory. If not provided, it is defaulted
to 'ROLE_USER'. It creates a user according to the parameters, put it in the database, and flush the data. In the future, if necessary, it could be useful
to add a parameter to ask if we want to flush the database, because it could maybe be slow if we need to do that a lot.