:bug: (Issue) fixed issue 32
when we were making a request on PATCH /users/{id}, we couldn't pass the "address" tag without the server producing an error. When the old addresses were
removed, they were not deleted, it was only the link which was. So I simply set the parameter orphanRemoval to true in the call to the annotation OneToMany
on top of the address field in the User entity. Now, when an address is removed, it is also deleted from the database

Closes #32