:rewind: (tests and Address) commited wrong files
I commited 2 test classes in my last commit, which should not be in this branch. So I removed them. I also commited a
change in the UserAddress entity, which seems to not affect anything. I discovered too that I changed the
UserAdressRepository. This change was not reverted because UserAddress was written with only one "d", which was an
error.