:sparkles: (Routes /ues) Added routes for UEs
Added route GET /ues/search : it can search UEs by their code and the type of ECTS they give. Added route GET /ues/{code} that returns informations about the UE with the provided code. Added a new search
filter for ApiPlatform which allows to change the name of the parameters in the URL. Fixed a small bug in the UpdateAtSubscriber class that could make the server produce an error