feat: First part of the exposition of the Asso, Event, and UE entities (#24)
* feat: First part of the exposition of the Asso Event and UE entities

* :bug: (Asso) Fixed a few problems of PR#24

createdAt and keywords are now exposed. Fixed import errors. Linted a space. Merged both class annotations in the same #[]. Added a sort based on the asso name. Added a filter on keywords

* :bug: (Routes /assos) Polished the PR

null values will now be returned in routes /assos, /events and /ues. Descriptions are now returned in route GET /assos (short description) and GET /assos/{id} (long description). Memberships and
membership count are now exposed.

* :bug: (/ues and /events) These files should have been in the previous PR. null values were not sent

set attribute "skip_null_values" to false in the exposition of /ues and /events

* :bug: (/ues, /events) Polishing

Added security for /ues and /events routes. Parameters of ApiResource are now in the right order for files UE, Event and Asso. Changed filter for UE searching from exact to partial

* :arrow_up: (PHP) upgrading from PHP 8.0 to PHP 8.1

This may be the reason why build are failing in PR#24

* :bug: (php version) changed php version for tests

* :bug: (composer.lock) regenerated composer.lock

* :rotating_light: (Asso.php and Event.php) Linted the whole project

Co-authored-by: pierre <pierre.nicholson@utt.fr>
Co-authored-by: Teddy Roncin <teddy.roncin35@gmail.com>